move docs script
