+++ scripts/docs.js
@@ -2,6 +2,7 @@ const fs = require('fs')
 const path = require('path')
 const markdownMagic = require('markdown-magic')
 const globby = require('markdown-magic').globby
+const generateCommandData = require('./generateCommandData')
 
 process.env.DOCS_GEN = 'TRUE'
 
@@ -74,14 +75,6 @@ markdownMagic(['README.md', 'docs/**/**.md'], config, () => {
   })
   console.log('Docs updated!')
 })
-
-function commandFromPath(p) {
-  return p.replace(process.cwd(), '')
-    .replace('.js', '')
-    .replace('/src/commands/', '')
-    .replace('/index', '')
-    .replace('/', ':')
-}
 
 /* Start - Docs Templating logic */
 function commandExamples(examples) {
@@ -179,64 +172,3 @@ function formatArgs(cmdArgs) {
   return renderArgs
 }
 /* End - Docs Templating logic */
-
-function generateCommandData() {
-  const commandsPath = path.join(__dirname, 'src/commands')
-  const commands = globby.sync([`${commandsPath}/**/**.js`])
-
-  const allCommands = commands.map((file) => {
-    let cmd = {}
-    try {
-      cmd = require(file)
-    } catch (e) {
-      throw e
-    }
-    const command = commandFromPath(file)
-    const parentCommand = command.split(':')[0]
-    const parent = (command === parentCommand) ? true : false
-    return {
-      command: command,
-      commandGroup: parentCommand,
-      isParent: parent,
-      path: file,
-      data: cmd
-    }
-  })
-
-  const visibleCommands = allCommands.filter((cmd) => {
-    return !cmd.data.hidden
-  })
-
-  console.log('visibleCommands', visibleCommands)
-
-  const groupedCommands = visibleCommands.reduce((acc, curr) => {
-    if (curr.commandGroup === curr.command) {
-      acc[curr.commandGroup] = {
-        name: curr.command,
-        description: curr.data.description,
-        flags: curr.data.flags,
-        args: curr.data.args,
-        examples: curr.data.examples,
-        strict: curr.data.strict,
-        commands: []
-      }
-    }
-    return acc
-  }, {})
-
-  const groupedCommandsWithData = visibleCommands.reduce((acc, curr) => {
-    if (curr.commandGroup !== curr.command) {
-      acc[curr.commandGroup].commands = acc[curr.commandGroup].commands.concat({
-        name: curr.command,
-        description: curr.data.description,
-        flags: curr.data.flags,
-        args: curr.data.args,
-        examples: curr.data.examples,
-        strict: curr.data.strict
-      })
-    }
-    return acc
-  }, groupedCommands)
-
-  return groupedCommandsWithData
-}
